+++ DepositCalculator/Calculator/InterestCalculatorBase.cs
@@ -2,6 +2,6 @@
 {
     public abstract class InterestCalculatorBase
     {
-        public abstract double Calc(double amount, int termMonthes, double rate, out IEnumerable<CalcResult> calculationDetails);
+        public abstract decimal Calc(decimal amount, int termMonthes, decimal rate, out IEnumerable<CalcResult> calculationDetails);
     }
 }
